+++ EliteAPI/Events/Location.cs
@@ -38,7 +38,7 @@ namespace EliteAPI
         public string SystemGovernment_Localised { get; set; }
         public string SystemSecurity { get; set; }
         public string SystemSecurity_Localised { get; set; }
-        public int Population { get; set; }
+        public ulong Population { get; set; }
         public string Body { get; set; }
         public int BodyID { get; set; }
         public string BodyType { get; set; }
